Extract unit test for logic moved out of service
